Fixed deprecation warnings
Deprecation warnings prior:

WARN 2020/03/30 18:12:29 Page.Hugo is deprecated and will be removed in a future release. Use the global hugo function.
WARN 2020/03/30 18:12:29 Page.URL is deprecated and will be removed in a future release. Use .Permalink or .RelPermalink. If what you want is the front matter URL value, use .Params.url
WARN 2020/03/30 18:12:29 Page.UniqueID is deprecated and will be removed in a future release. Use .File.UniqueID
WARN 2020/03/30 18:12:29 Page.Dir is deprecated and will be removed in a future release. Use .File.Dir
WARN 2020/03/30 18:12:29 .File.UniqueID on zero object. Wrap it in if or with: {{ with .File }}{{ .UniqueID }}{{ end }}